// examples/mock-http-server/src/types/generated.ts
export type CartOperation =
  /**
   * A list of valid discount codes that correspond to external discounts. This can
   * only be used by Functions with network access.
   */
  | {
      enteredDiscountCodesAccept: ValidDiscountCodes;
      orderDiscountsAdd?: never;
      productDiscountsAdd?: never;
    } /** A group of order discounts that share a selection strategy. */
  | {
      enteredDiscountCodesAccept?: never;
      orderDiscountsAdd: OrderDiscounts;
      productDiscountsAdd?: never;
    } /** A group of product discounts that share a selection strategy. */
  | {
      enteredDiscountCodesAccept?: never;
      orderDiscountsAdd?: never;
      productDiscountsAdd: ProductDiscounts;
    };
// ...
export type DeliveryDiscountCandidate = {
  /** The discount code associated with this discount candidate, for code-based discounts. */
  associatedDiscountCode?: InputMaybe<AssociatedDiscountCode>;
  /** The discount message. */
  message?: InputMaybe<Scalars["String"]["input"]>;
  /** The targets of the delivery discount candidate. */
  targets: Array<DeliveryDiscountCandidateTarget>;
  /** The value of the delivery discount candidate. */
  value: DeliveryDiscountCandidateValue;
};
// ...
export enum DeliveryDiscountSelectionStrategy {
  /**
   * Apply all delivery discount candidates with conditions that are satisfied. This does not override
   * discount combination or stacking rules.
   */
  All = "ALL",
}

/** A group of delivery discounts that share a selection strategy. */
export type DeliveryDiscounts = {
  /** The list of delivery discount candidates to be applied. */
  candidates: Array<DeliveryDiscountCandidate>;
  /** The strategy that's applied to the list of discounts. */
  selectionStrategy: DeliveryDiscountSelectionStrategy;
};
// ...
export type DeliveryOperation =
  /** A group of delivery discounts that share a selection strategy. */
  | {
      deliveryDiscountsAdd: DeliveryDiscounts;
      enteredDiscountCodesAccept?: never;
    } /**
   * A list of valid discount codes that correspond to external discounts. This can
   * only be used by Functions with network access.
   */
  | {
      deliveryDiscountsAdd?: never;
      enteredDiscountCodesAccept: ValidDiscountCodes;
    };
// ...
export type OrderDiscountCandidate = {
  /** The discount code associated with this discount candidate, for code-based discounts. */
  associatedDiscountCode?: InputMaybe<AssociatedDiscountCode>;
  /** The conditions that must be satisfied to apply the order discount candidate. */
  conditions?: InputMaybe<Array<Condition>>;
  /** The discount message. */
  message?: InputMaybe<Scalars["String"]["input"]>;
  /** The targets of the order discount candidate. */
  targets: Array<OrderDiscountCandidateTarget>;
  /** The value of the order discount candidate. */
  value: OrderDiscountCandidateValue;
};
// ...
export enum OrderDiscountSelectionStrategy {
  /** Only apply the first order discount candidate with conditions that are satisfied. */
  First = "FIRST",
  /** Only apply the order discount candidate that offers the maximum reduction. */
  Maximum = "MAXIMUM",
}

/** A group of order discounts that share a selection strategy. */
export type OrderDiscounts = {
  /** The list of order discount candidates to be applied. */
  candidates: Array<OrderDiscountCandidate>;
  /** The strategy that's applied to the list of discounts. */
  selectionStrategy: OrderDiscountSelectionStrategy;
};
// ...
export type ProductDiscountCandidate = {
  /** The discount code associated with this discount candidate, for code-based discounts. */
  associatedDiscountCode?: InputMaybe<AssociatedDiscountCode>;
  /** The discount message. */
  message?: InputMaybe<Scalars["String"]["input"]>;
  /** The targets of the product discount candidate. */
  targets: Array<ProductDiscountCandidateTarget>;
  /** The value of the product discount candidate. */
  value: ProductDiscountCandidateValue;
};
// ...
export enum ProductDiscountSelectionStrategy {
  /**
   * Apply all product discount candidates with conditions that are satisfied. This
   * does not override discount combination or stacking rules.
   */
  All = "ALL",
  /** Only apply the first product discount candidate with conditions that are satisfied. */
  First = "FIRST",
  /** Only apply the product discount candidate that offers the maximum reduction. */
  Maximum = "MAXIMUM",
}

/** A group of product discounts that share a selection strategy. */
export type ProductDiscounts = {
  /** The list of product discount candidates to be applied. */
  candidates: Array<ProductDiscountCandidate>;
  /** The strategy that's applied to the list of product discount candidates. */
  selectionStrategy: ProductDiscountSelectionStrategy;
};
// ...
export type ValidDiscountCode = {
  /** The discount code. */
  code: Scalars["String"]["input"];
};

/**
 * A list of valid discount codes that correspond to external discounts. This can
 * only be used by Functions with network access.
 */
export type ValidDiscountCodes = {
  /** The list of valid discount codes. */
  codes: Array<ValidDiscountCode>;
};